Update the old php config cleaning

+++ install/rhel/10/templates/web/nginx/php82.sh
@@ -11,9 +11,15 @@ chmod 0700  "$home_dir/$user/web/$domain/tmp"
 chown "$user":"$user" "$home_dir/$user/web/$domain/tmp"
 
 phpv=82
-oldphpv="$(ls /etc/opt/remi/php*/php-fpm.d/${user}-${domain}.conf | awk -F'/' '{print $5}')"
-rm -f "/etc/opt/remi/${oldphpv}/php-fpm.d/${user}-${domain}.conf"
-sudo systemctl reload "${oldphpv}-php-fpm" || sudo systemctl start "${oldphpv}-php-fpm"
+if ls /etc/opt/remi/php*/php-fpm.d/"${user}-${domain}.conf" ; then
+  oldphpvs="$(ls /etc/opt/remi/php*/php-fpm.d/"${user}-${domain}.conf" | awk -F'/' '{print $5}')"
+  for oldphpv in $oldphpvs ; do
+    if [[ "$oldphpv" != "$phpv" ]] ; then
+      rm -f "/etc/opt/remi/${oldphpv}/php-fpm.d/${user}-${domain}.conf"
+      sudo systemctl reload "${oldphpv}-php-fpm" || sudo systemctl start "${oldphpv}-php-fpm"
+    fi
+  done
+fi
 cat > "/etc/opt/remi/php$phpv/php-fpm.d/${user}-${domain}.conf" << EOF
 [${user}-${domain}]
 user = $user
